microservices/dynamicPricing/cloudFunction/requirements.txt
functions-framework==3.*
pymongo
scikit-learn==1.2.2
numpy
google-cloud-aiplatform
google-auth
google-cloud-storage
pandas
joblib